+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 [Next release](https://github.com/rebing/graphql-laravel/compare/v1.24.0...master)
 ------------
 ## Breaking changes
+- The `UploadType` now has to be added manually to the `types` in your schema if you want to use it
+  - The `::getInstance()` method is gone
 - The order and arguments/types for resolvers has changed:
   - before: `resolve($root, $array, SelectFields $selectFields, ResolveInfo $info)`
   - after: `resolve($root, $array, $context, ResolveInfo $info, Closure $getSelectFields)`